refactor out header for C runtime functions
